feat: update data models for booking requests and user information, enhance error handling in dashboard socket repository, and improve dashboard UI with state management

+++ lib/features/dashboard/data/models/Analysis_model/request_guide_model/booking_request_model.dart
@@ -7,7 +7,7 @@ class BookingRequestModel {
   final PlaceModel place;
   final DateTime bookingDate;
   final int persons;
-  final double totalPrice;
+  final int totalPrice;
   final String status;
 
   BookingRequestModel({
